Move parameterized test extension and utils in iceberg-api
